Move ServiceResources to service.go.
